+++ src/i18n/id.extra.js
@@ -1,0 +1,6 @@
+window.ParsleyConfig = window.ParsleyConfig || {};
+window.ParsleyConfig.i18n = window.ParsleyConfig.i18n || {};
+
+window.ParsleyConfig.i18n.id = $.extend(window.ParsleyConfig.i18n.id || {}, {
+  dateiso: "Harus tanggal yang valid (YYYY-MM-DD)."
+});