+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Parsley 2.x changelog
 
+## 2.0.1
+
+  - fixed "attr.specified is deprecated." console warning (#608)
+  - fixed package.json config with main dep (#617)
+  - fixed `addValidator()` method
+  - added support for `requirementsTransformer` for custom validators
+  - updated jQuery needed version from README and doc
+  - fixed case when a multiple item were dynamically removed from DOM (#634)
+  - added proper `type="range"` support (#668)
+
 ## 2.0.0
 
   - fixed remote re-entering already validated value (#576)